more correction

+++ toem_utils.c
@@ -1,26 +1,44 @@
 #include "shell.h"
 
+/**
+ * _strlen - Calculate the length of a string.
+ * @str: The input string.
+ *
+ * Return: The length of the string.
+ */
 size_t _strlen(const char *str)
 {
-    int len = 0;
+    size_t len = 0;
     while (str[len] != '\0')
         len++;
     return len;
 }
 
+/**
+ * _strdup - Duplicate a string.
+ * @src: The source string to duplicate.
+ *
+ * Return: A pointer to the duplicated string, or NULL on failure.
+ */
 char *_strdup(const char *src)
 {
-    int len = _strlen(src);
+    size_t len = _strlen(src);
     char *dup = malloc(len + 1);
     if (dup)
         return memcpy(dup, src, len + 1);
     return NULL;
 }
 
+/**
+ * free_info - Free the memory allocated for an info_t structure.
+ * @info: Pointer to the info_t structure to free.
+ * @free_env: Flag to indicate whether to free the env array.
+ */
 void free_info(info_t *info, int free_env)
 {
     if (info->arg)
         free(info->arg);
+
     if (free_env && info->env)
     {
         char **env = info->env;
@@ -35,12 +53,12 @@ void free_info(info_t *info, int free_env)
 }
 
 /**
- * _strncmp - Compare two strings up to n bytes
- * @s1: The first string
- * @s2: The second string
- * @n: The maximum number of bytes to compare
+ * _strncmp - Compare two strings up to n bytes.
+ * @s1: The first string.
+ * @s2: The second string.
+ * @n: The maximum number of bytes to compare.
  *
- * Return: 0 if strings are equal, < 0 if s1 < s2, > 0 if s1 > s2
+ * Return: 0 if strings are equal, < 0 if s1 < s2, > 0 if s1 > s2.
  */
 int _strncmp(const char *s1, const char *s2, size_t n)
 {
